Add subCcdGray functionality, and other fixes.
Add in chebyshev polynomial fits of ccdGray, and combine with the chebyshev
fits of the superstar flats if required.  The chebyshev field code has been
overhauled to conform with the lsst stack (including a 0.5 pixel offset).
Zeropoints are only output if desired to save disk space.  Finally, a fix has
been put in to limit the number of numpy threads for being a better neighbor on
compute clusters.