Update code with better input checking and outputs.  Version now 2.4.0
Input magnitudes are now sanitized and flagged if they are bad, so that nans
don't spoil everything.  The input airmass of all observations is checked
to confirm that they are within range of the LUT.  Otherwise, the LUT code
will not work properly.  When "fgcmcal" (lsst) output is selected all
zeropoints are possibly variable and flux-based, allowing better composition
of constituent parts.